* 'master' of https://github.com/webvimark/user-management:
  Label Login correctly translated
  Fix migration if cache not present
  update readme

+++ models/rbacDB/Route.php
@@ -142,7 +142,9 @@ class Route extends AbstractItem
 
 		if ( $toAdd || $toRemove )
 		{
-			Yii::$app->cache->delete('__commonRoutes');
+			if (Yii::$app->cache) {
+				Yii::$app->cache->delete('__commonRoutes');
+			}
 		}
 	}
